+++ src/adapters/index.ts
@@ -1,2 +1,2 @@
-export { init_express } from "./express";
 export { create_openapi_schema } from "./openapi";
+export { nodeServerHandle } from "./node";