+++ src/contracts/core.ts
@@ -1,25 +1,13 @@
-import { Readable } from "stream";
+import type { Readable } from "stream";
+import type { IncomingMessage, ServerResponse } from "http";
 
 export type IOpenRequestBody = Record<string, string | number | boolean | IOpenRequestFile>;
 
-export interface OpenRequest {
-  readonly params: Record<string, string>;
-  readonly query: Record<string, string>;
-  readonly headers: Record<string, string>;
+export interface OpenRequest extends IncomingMessage {
   readonly body: IOpenRequestBody;
-  readonly secure: boolean;
-  readonly path: string;
-  readonly hostname: string;
-  readonly method: string;
 }
 
-export interface OpenResponse {
-  setHeader(name: string, value: string): any;
-  statusCode: number;
-  writableEnded: boolean;
-  write(data: string): any;
-  end(data?: string): any;
-}
+export interface OpenResponse extends ServerResponse {}
 
 export interface IOpenRequestFile {
   filename: string;