+++ test/Route.test.ts
@@ -1,11 +1,10 @@
-import { describe, it, expect, mock } from "bun:test";
-import { RouteBuilder, Route } from "../src/Route";
 import * as t from "@sinclair/typebox";
+import { describe, expect, it, mock } from "bun:test";
+import { Route, RouteBuilder } from "../src/Route";
 
 const createMockResponse = () => ({
   writableEnded: false,
-  setHeader: mock(() => null),
-  statusCode: 200,
+  writeHead: mock((_, __) => null),
   write: mock((data) => data),
   end: mock((data) => data),
 });
@@ -18,56 +17,56 @@ describe("Route", () => {
 
   it("create GET /articles route", () => {
     const created = Route()
-      .setPath("get", "/articles")
+      .setPath("GET", "/articles")
       .setHandle(() => {})
       .build();
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("get");
+    expect(created.method).toBe("GET");
     expect(created.__handle).toBeFunction();
   });
 
   it("create POST /article route", () => {
     const created = Route()
-      .setPath("post", "/articles")
+      .setPath("POST", "/articles")
       .setHandle(() => {})
       .build();
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("post");
+    expect(created.method).toBe("POST");
     expect(created.__handle).toBeFunction();
   });
 
   it("create PUT /article route", () => {
     const created = Route()
-      .setPath("put", "/articles")
+      .setPath("PUT", "/articles")
       .setHandle(() => {})
       .build();
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("put");
+    expect(created.method).toBe("PUT");
     expect(created.__handle).toBeFunction();
   });
 
   it("create PATCH /article route", () => {
     const created = Route()
-      .setPath("patch", "/articles")
+      .setPath("PATCH", "/articles")
       .setHandle(() => {})
       .build();
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("patch");
+    expect(created.method).toBe("PATCH");
     expect(created.__handle).toBeFunction();
   });
 
   it("create DELETE /article route", () => {
     const created = Route()
-      .setPath("delete", "/articles")
+      .setPath("DELETE", "/articles")
       .setHandle(() => {})
       .build();
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("delete");
+    expect(created.method).toBe("DELETE");
     expect(created.__handle).toBeFunction();
   });
 
@@ -75,7 +74,7 @@ describe("Route", () => {
     const middleware1 = mock(async () => {});
     const middleware2 = mock(async () => {});
     const created = Route()
-      .setPath("get", "/articles")
+      .setPath("GET", "/articles")
       .setMiddlewares(middleware1, middleware2)
       .setHandle(() => {})
       .build();
@@ -84,7 +83,7 @@ describe("Route", () => {
     await created.__handle({} as any, response);
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("get");
+    expect(created.method).toBe("GET");
     expect(created.__handle).toBeFunction();
     expect(middleware1).toHaveBeenCalled();
     expect(middleware2).toHaveBeenCalled();
@@ -92,32 +91,32 @@ describe("Route", () => {
 
   it("supports handle for GET /articles", async () => {
     const handle = mock(async () => {});
-    const created = Route().setPath("get", "/articles").setHandle(handle).build();
+    const created = Route().setPath("GET", "/articles").setHandle(handle).build();
 
     const response = createMockResponse();
     await created.__handle({} as any, response);
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("get");
+    expect(created.method).toBe("GET");
     expect(created.__handle).toBeFunction();
     expect(handle).toHaveBeenCalled();
   });
 
   it("supports params for /articles/:id", () => {
     const created = Route()
-      .setPath("get", "/articles/:id")
+      .setPath("GET", "/articles/:id")
       .setParams(t.Object({ id: t.String({}) }))
       .setHandle(() => {})
       .build();
 
     expect(created.path).toBe("/articles/:id");
-    expect(created.method).toBe("get");
+    expect(created.method).toBe("GET");
     expect(created.__handle).toBeFunction();
   });
 
   it("supports params for /articles/:article_id/comments/:comment_id", () => {
     const created = Route()
-      .setPath("get", "/articles/:article_id/comments/:comment_id")
+      .setPath("GET", "/articles/:article_id/comments/:comment_id")
       .setParams(
         t.Object({
           article_id: t.String(),
@@ -128,14 +127,14 @@ describe("Route", () => {
       .build();
 
     expect(created.path).toBe("/articles/:article_id/comments/:comment_id");
-    expect(created.method).toBe("get");
+    expect(created.method).toBe("GET");
     expect(created.__handle).toBeFunction();
   });
 
   it("validates params presence from route path /articles/:id", () => {
     const created = () =>
       Route()
-        .setPath("get", "/articles/:id")
+        .setPath("GET", "/articles/:id")
         .setHandle(() => {})
         .build();
 
@@ -154,18 +153,18 @@ describe("Route", () => {
 
   it("treats blank path for route as /", () => {
     const created = Route()
-      .setPath("get", "")
+      .setPath("GET", "")
       .setHandle(() => {})
       .build();
 
     expect(created.path).toBe("/");
-    expect(created.method).toBe("get");
+    expect(created.method).toBe("GET");
     expect(created.__handle).toBeFunction();
   });
 
   it("works for valid request data for POST /articles", async () => {
     const created = Route()
-      .setPath("post", "/articles")
+      .setPath("POST", "/articles")
       .setRequestData(
         t.Object({
           title: t.String({}),
@@ -179,14 +178,14 @@ describe("Route", () => {
     await created.__handle({ body: { title: "abc", description: "abcd" } } as any, response);
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("post");
+    expect(created.method).toBe("POST");
     expect(created.__handle).toBeFunction();
     expect(response.write).toHaveBeenCalled();
   });
 
   it("fails for invalid request data for POST /articles", async () => {
     const created = Route()
-      .setPath("post", "/articles")
+      .setPath("POST", "/articles")
       .setRequestData(
         t.Object({
           title: t.String(),
@@ -200,14 +199,14 @@ describe("Route", () => {
     await created.__handle({} as any, response);
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("post");
+    expect(created.method).toBe("POST");
     expect(created.__handle).toBeFunction();
     expect(response.write.mock.results[0].value).not.toBe("{}");
   });
 
   it("validates request without data for POST /articles", () => {
     const created = Route()
-      .setPath("post", "/articles")
+      .setPath("POST", "/articles")
       .setHandle(() => {})
       .build();
 
@@ -215,7 +214,7 @@ describe("Route", () => {
     const success = () => created.__handle({ body: {} } as any, response);
 
     expect(created.path).toBe("/articles");
-    expect(created.method).toBe("post");
+    expect(created.method).toBe("POST");
     expect(created.__handle).toBeFunction();
     expect(success).not.toThrowError();
   });
@@ -223,7 +222,7 @@ describe("Route", () => {
   it("disallows to use Route().setRequestData with GET and OPTIONS method", () => {
     const get = () =>
       Route()
-        .setPath("get", "/articles")
+        .setPath("GET", "/articles")
         .setRequestData(
           t.Object({
             title: t.String(),
@@ -235,7 +234,7 @@ describe("Route", () => {
 
     const options = () =>
       Route()
-        .setPath("options", "/articles")
+        .setPath("OPTIONS", "/articles")
         .setRequestData(
           t.Object({
             title: t.String(),